backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Data Validation
email-validator==2.1.0

# Caching & Task Queue
redis==5.0.1
celery==5.3.4

# HTTP Client
httpx==0.25.2
requests==2.31.0

# OAuth
authlib==1.3.0

# Email
resend==0.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Logging
python-json-logger==2.0.7

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# AI & ML
google-generativeai==0.3.0
